+++ playground/vue-server-origin/__tests__/vue-server-origin.spec.ts
@@ -3,7 +3,7 @@ import { isBuild, page } from '~utils'
 
 test('should render', async () => {
   const expected = isBuild
-    ? /assets\/asset\.[0-9a-f]+\.png/
+    ? /assets\/asset\.[\da-f]+\.png/
     : 'http://localhost/server-origin/test/assets/asset.png'
 
   expect(await page.getAttribute('img', 'src')).toMatch(expected)